#10-22 김승환 이동동기화 수정
x,z축 동기화 완료, y축은 수정 중.

+++ Texture_Version07-9-1/ObjectsInfo.h
@@ -6,7 +6,7 @@ enum OBJECT_STATE { ST_EMPTY, ST_RUNNING };
 struct ObjectsInfo
 {
 	short m_id;
-	short m_x, m_y, m_z;
+	float m_x, m_y, m_z;
 	int m_state;
 
 	ObjectsInfo() {
@@ -14,7 +14,7 @@ struct ObjectsInfo
 		m_x = 0, m_y = 0, m_z = 0;
 		m_state = ST_EMPTY;
 	}
-	ObjectsInfo(short id, short x, short y, short z) {
+	ObjectsInfo(short id, float x, float y, float z) {
 		m_id = id;
 		m_x = x, m_y = y, m_z = z;
 		m_state = ST_RUNNING;
